issues/requirements.txt
Jinja2
PyGithub
pyyaml
tabulate